requirements.txt
asgiref==3.8.1
Django==4.1.9
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1
